refactor: reorganize module structure and update import paths for consistency
