+++ src/modules/core/Cell.ts
@@ -1,8 +1,8 @@
 import * as PIXI from 'pixi.js';
-import EmptyField from '../assets/sprites/blocks/grass-tile.png';
-import EmptyFieldSecond from '../assets/sprites/blocks/grass-tile-second.png';
-import { GameObject } from './GameObject';
-import AvaibleCell from '../assets/sprites/blocks/second-availible.png';
+import EmptyField from '@/assets/sprites/blocks/grass-tile.png';
+import EmptyFieldSecond from '@/assets/sprites/blocks/grass-tile-second.png';
+import { GameObject } from '@/modules/core/GameObject';
+import AvaibleCell from '@/assets/sprites/blocks/second-availible.png';
 
 export default class Cell extends PIXI.Container {
   public sprite: PIXI.Sprite;
