+++ src/modules/core/Grid.ts
@@ -1,7 +1,7 @@
 import * as PIXI from 'pixi.js';
-import Cell from './Cell';
-import { GameObject } from './GameObject';
-import { Colors, getRandomColor } from './../utils';
+import Cell from '@/modules/core/Cell';
+import { GameObject } from '@/modules/core/GameObject';
+import { Colors, getRandomColor } from '@/utils';
 const DEFAULT_GRID_SIZE = 5;
 
 export default class Grid {
