+++ tsconfig.json
@@ -1,5 +1,8 @@
 {
   "compilerOptions": {
+    "paths": {
+      "@/*": ["./*"]
+    },
     "target": "ES2020",
     "useDefineForClassFields": true,
     "allowSyntheticDefaultImports": true,
