+++ vite.config.ts
@@ -1,3 +1,10 @@
+import path from 'path';
 import { defineConfig } from 'vite';
 
-export default defineConfig({});
+export default defineConfig({
+  resolve: {
+    alias: {
+      '@': path.resolve(__dirname, 'src'),
+    },
+  },
+});
